requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
jupyter>=1.0.0
openai>=1.0.0
google-generativeai>=0.3.0
